Openwrt : Patch to enable ubus on server side.
Issue : Currently we are not able to do UCI SET from server code of
qcmap as it is running in radio

Fix :Enable ubus on server side to so, that we can
do UCI SET from server code.
But ubox and ubus lib's are in c. Require this patch to include
c lib from cpp code.

From: Dheeraj Kumar <quic_dhekum@quicinc.com>

Change-Id: If1add3b3db278380d8084741ba385f54dd8f8bcd