odhcpd: solve race condition for ndp
odhcpd attempts to re-initialize a socket on an interface that the kernel
is simultaneously tearing down.
Set the callback pointer to NULL before deleting the file descriptor from the uloop,
which can prevent the callback from executing on a closed file descriptor
Gracefully handle Bad file descriptor in odhcpd_receive_packets to avoid
infinite loop.

Change-Id: Ia3d36c3153e4e1b40168dbe7566a8c8f254e20c4